requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
requests==2.31.0
beautifulsoup4==4.12.3
sentence-transformers==2.5.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
openai==1.14.2
google-generativeai==0.4.1
pytest==8.1.1
httpx==0.27.0
